+++ src/main/java/com/fyrdev/monyia/adapters/driving/http/dto/request/GoalTransactionRequest.java
@@ -2,5 +2,7 @@ package com.fyrdev.monyia.adapters.driving.http.dto.request;
 
 import com.fyrdev.monyia.domain.model.enums.GoalTransactionType;
 
-public record GoalTransactionRequest(Long goalId, Double amount, GoalTransactionType type) {
+import java.math.BigDecimal;
+
+public record GoalTransactionRequest(Long goalId, BigDecimal amount, GoalTransactionType type) {
 }
