+++ src/main/java/com/fyrdev/monyia/adapters/driving/http/dto/request/TransferRequest.java
@@ -1,4 +1,6 @@
 package com.fyrdev.monyia.adapters.driving.http.dto.request;
 
-public record TransferRequest(Long fromPocketId, Long toPocketId, Double amount) {
+import java.math.BigDecimal;
+
+public record TransferRequest(Long fromPocketId, Long toPocketId, BigDecimal amount) {
 }
